# namviet-backend-core/docs/swagger.yaml
definitions:
...
  medical_visits.CreateMedicalVisitRequest:
    properties:
      appointment_id:
        type: string
      customer_id:
        type: integer
      doctor_id:
        type: string
      symptoms:
        type: string
    required:
    - customer_id
    type: object
  medical_visits.MedicalVisit:
    properties:
      appointment_id:
        type: string
      bp_diastolic:
        type: integer
      bp_systolic:
        type: integer
      created_at:
        type: string
      created_by:
        type: string
      customer_id:
        type: integer
      deleted_at:
        type: string
      diagnosis:
        type: string
      doctor_id:
        type: string
      doctor_notes:
        type: string
      examination_summary:
        type: string
      height:
        type: number
      icd_code:
        type: string
      id:
        type: string
      pulse:
        type: integer
      red_flags:
        items:
          type: string
        type: array
      sp02:
        type: integer
      status:
        type: string
      symptoms:
        type: string
      temperature:
        type: number
      updated_at:
        type: string
      updated_by:
        type: string
      weight:
        type: number
    type: object
  medical_visits.UpdateMedicalVisitRequest:
    properties:
      bp_diastolic:
        type: integer
      bp_systolic:
        type: integer
      diagnosis:
        type: string
      doctor_notes:
        type: string
      examination_summary:
        type: string
      height:
        type: number
      icd_code:
        type: string
      pulse:
        type: integer
      red_flags:
        items:
          type: string
        type: array
      sp02:
        type: integer
      status:
        type: string
      symptoms:
        type: string
      temperature:
        type: number
      weight:
        type: number
    type: object
...
  roles.CreateRoleRequest:
    properties:
      description:
        type: string
      name:
        type: string
      permissions:
        items:
          type: string
        type: array
    required:
    - name
    type: object
  roles.Role:
    properties:
      created_at:
        type: string
      description:
        type: string
      id:
        type: string
      name:
        type: string
      permissions:
        items:
          type: string
        type: array
    type: object
  roles.UpdateRoleRequest:
    properties:
      description:
        type: string
      name:
        type: string
      permissions:
        items:
          type: string
        type: array
    type: object
...
info:
  contact: {}
  description: Backend API for Nam Viet ERP System
  title: Nam Viet ERP API
  version: "1.0"
paths:
...
  /medical-visits:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all medical visits
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/medical_visits.MedicalVisit'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get All Medical Visits
      tags:
      - Medical Visits
    post:
      consumes:
      - application/json
      description: Create a new medical visit
      parameters:
      - description: Medical Visit Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/medical_visits.CreateMedicalVisitRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/medical_visits.MedicalVisit'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new Medical Visit
      tags:
      - Medical Visits
  /medical-visits/{id}:
    delete:
      consumes:
      - application/json
      description: Soft delete a medical visit
      parameters:
      - description: Medical Visit ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete Medical Visit
      tags:
      - Medical Visits
    get:
      consumes:
      - application/json
      description: Retrieve a specific medical visit
      parameters:
      - description: Medical Visit ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/medical_visits.MedicalVisit'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get Medical Visit by ID
      tags:
      - Medical Visits
    put:
      consumes:
      - application/json
      description: Update medical visit details
      parameters:
      - description: Medical Visit ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/medical_visits.UpdateMedicalVisitRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/medical_visits.MedicalVisit'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update Medical Visit
      tags:
      - Medical Visits
...
  /roles:
    get:
      consumes:
      - application/json
      description: Retrieve a list of all roles
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            items:
              $ref: '#/definitions/roles.Role'
            type: array
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get All Roles
      tags:
      - Roles
    post:
      consumes:
      - application/json
      description: Create a new role with permissions
      parameters:
      - description: Role Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/roles.CreateRoleRequest'
      produces:
      - application/json
      responses:
        "201":
          description: Created
          schema:
            $ref: '#/definitions/roles.Role'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Create a new Role
      tags:
      - Roles
  /roles/{id}:
    delete:
      consumes:
      - application/json
      description: Delete a role
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "204":
          description: No Content
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Delete Role
      tags:
      - Roles
    get:
      consumes:
      - application/json
      description: Retrieve a specific role
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/roles.Role'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Get Role by ID
      tags:
      - Roles
    put:
      consumes:
      - application/json
      description: Update role details and permissions
      parameters:
      - description: Role ID
        in: path
        name: id
        required: true
        type: string
      - description: Update Details
        in: body
        name: request
        required: true
        schema:
          $ref: '#/definitions/roles.UpdateRoleRequest'
      produces:
      - application/json
      responses:
        "200":
          description: OK
          schema:
            $ref: '#/definitions/roles.Role'
        "400":
          description: Bad Request
          schema:
            additionalProperties:
              type: string
            type: object
        "401":
          description: Unauthorized
          schema:
            additionalProperties:
              type: string
            type: object
        "404":
          description: Not Found
          schema:
            additionalProperties:
              type: string
            type: object
        "500":
          description: Internal Server Error
          schema:
            additionalProperties:
              type: string
            type: object
      security:
      - BearerAuth: []
      summary: Update Role
      tags:
      - Roles
...
securityDefinitions:
  BearerAuth:
    in: header
    name: Authorization
    type: apiKey
swagger: "2.0"